100AP project integration
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@2402 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Animation Control.h
@@ -556,7 +556,7 @@ extern ANI_SPEED_DEF gubAnimWalkSpeeds[ TOTALBODYTYPES ];
 extern ANI_SPEED_DEF gubAnimRunSpeeds[ TOTALBODYTYPES ];
 extern ANI_SPEED_DEF gubAnimSwatSpeeds[ TOTALBODYTYPES ];
 extern ANI_SPEED_DEF gubAnimCrawlSpeeds[ TOTALBODYTYPES ];
-extern UINT8 gubMaxActionPoints[ TOTALBODYTYPES ];
+extern INT16 gubMaxActionPoints[ TOTALBODYTYPES ];
 
 
 // OK, this array contains definitions for random animations based on bodytype, total # allowed, and what is in their hand....
